% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbConnect,DrillDriver-method}
\alias{dbConnect,DrillDriver-method}
\title{Drill dbConnect}
\usage{
\S4method{dbConnect}{DrillDriver}(drv = "JDBCDriver",
  nodes = "localhost:2181", cluster_id = NULL, schema = NULL,
  use_zk = TRUE, ...)
}
\description{
Drill dbConnect
}

