% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbDataType,DrillJDBCConnection-method}
\alias{dbDataType,DrillJDBCConnection-method}
\title{Drill JDBC dbDataType}
\usage{
\S4method{dbDataType}{DrillJDBCConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \code{\linkS4class{DrillJDBCDriver}} object}

\item{obj}{Any R object}

\item{...}{Extra optional parameters}
}
\description{
Drill JDBC dbDataType
}
