% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\name{dbDisconnect,DrillConnection-method}
\alias{dbDisconnect,DrillConnection-method}
\title{Disconnect from Drill}
\usage{
\S4method{dbDisconnect}{DrillConnection}(conn, ...)
}
\description{
Disconnect from Drill
}
\keyword{internal}
