% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{methods}
\name{dbGetInfo,DrillDriver-method}
\alias{dbGetInfo,DrillDriver-method}
\alias{dbGetInfo,DrillConnection-method}
\title{Metadata about database objects}
\usage{
\S4method{dbGetInfo}{DrillDriver}(dbObj)

\S4method{dbGetInfo}{DrillConnection}(dbObj)
}
\arguments{
\item{dbObj}{A \code{\linkS4class{DrillDriver}} or \code{\linkS4class{DrillConnection}} object}
}
\description{
Metadata about database objects
}
