% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{methods}
\name{dbUnloadDriver,DrillDriver-method}
\alias{dbUnloadDriver,DrillDriver-method}
\title{Unload driver}
\usage{
\S4method{dbUnloadDriver}{DrillDriver}(drv, ...)
}
\arguments{
\item{drv}{driver}

\item{...}{Extra optional parameters}
}
\description{
Unload driver
}

