% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{methods}
\name{Drill}
\alias{Drill}
\alias{dbConnect,DrillDriver-method}
\title{Drill}
\usage{
Drill()

\S4method{dbConnect}{DrillDriver}(drv, host = "localhost", port = 8047L,
  ssl = FALSE, ...)
}
\arguments{
\item{drv}{An object created by \code{Drill()}}

\item{host}{host}

\item{port}{port}

\item{ssl}{use ssl?}

\item{...}{Extra optional parameters}
}
\description{
Drill

Connect to Drill
}

