% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_active}
\alias{drill_active}
\title{Test whether Drill HTTP REST API server is up}
\usage{
drill_active(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
This is a very simple test (performs \code{HEAD /} on \code{drill_server}
}
\examples{
\dontrun{
drill_active()
}
}

