% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_cancel}
\alias{drill_cancel}
\title{Cancel the query that has the given queryid}
\usage{
drill_cancel(drill_con, query_id)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{query_id}{the UUID of the query in standard UUID format that Drill assigns to each query.}
}
\description{
Cancel the query that has the given queryid
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
