% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\name{db_data_type.DrillJDBCConnection}
\alias{db_data_type.DrillJDBCConnection}
\alias{sql_escape_ident.DrillJDBCConnection}
\alias{sql_translate_env.DrillJDBCConnection}
\alias{src_tbls.src_dbi}
\title{Drill internals}
\usage{
\method{db_data_type}{DrillJDBCConnection}(con, fields, ...)

\method{sql_escape_ident}{DrillJDBCConnection}(con, x)

\method{sql_translate_env}{DrillJDBCConnection}(con)

\method{src_tbls}{src_dbi}(x)
}
\arguments{
\item{x}{x}
}
\description{
"SHOW DATABASES"
}
\keyword{internal}
