% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_metrics}
\alias{drill_metrics}
\title{Get the current memory metrics}
\usage{
drill_metrics(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Get the current memory metrics
}
\examples{
\dontrun{
drill_connection() \%>\% drill_metrics()
}
}
