% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_options}
\alias{drill_options}
\title{List the name, default, and data type of the system and session options}
\usage{
drill_options(drill_server = Sys.getenv("DRILL_URL"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
List the name, default, and data type of the system and session options
}

