% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_profile}
\alias{drill_profile}
\title{Get the profile of the query that has the given queryid.}
\usage{
drill_profile(query_id, drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get the profile of the query that has the given queryid.
}

