% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_profiles}
\alias{drill_profiles}
\title{Get the profiles of running and completed queries}
\usage{
drill_profiles(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Get the profiles of running and completed queries
}
\examples{
\dontrun{
drill_connection() \%>\% drill_profiles()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

