% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_set}
\alias{drill_set}
\title{Set Drill SYSTEM or SESSION options}
\usage{
drill_set(..., type = c("session", "system"),
  drill_server = Sys.getenv("DRILL_URL", unset = "http://localhost:8047"))
}
\arguments{
\item{...}{named parameters to be sent to \code{ALTER [SYSTEM|SESSION]}}

\item{type}{set the \code{session} or \code{system} parameter}

\item{drill_server}{base URL of the \code{drill} server}
}
\value{
a \code{tbl} (invisibly) with the \code{ALTER} queries sent and results, including errors.
}
\description{
Helper function to make it more R-like to set Drill SESSION or SYSTEM optons. It
handles the conversion of R types (like \code{TRUE}) to SQL types and automatically
quotes parameter values (when necessary).
}
\details{
If any query errors result, error messages will be presented to the console.
}
\examples{
\dontrun{
drill_set(exec.errors.verbose=TRUE, store.format="parquet", web.logs.max_lines=20000)
}
}

