% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_setting_reset}
\alias{drill_setting_reset}
\title{Changes (optionally, all) session settings back to system defaults}
\usage{
drill_setting_reset(..., drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{...}{bare name of system options to reset}

\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Changes (optionally, all) session settings back to system defaults
}

