% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_stats}
\alias{drill_stats}
\title{Get Drillbit information, such as ports numbers}
\usage{
drill_stats(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get Drillbit information, such as ports numbers
}
\examples{
\dontrun{
drill_stats()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

