% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_status}
\alias{drill_status}
\title{Get the status of Drill}
\usage{
drill_status(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get the status of Drill
}
\note{
The output of this is in a "viewer" window
}
\examples{
\dontrun{
drill_status()
}
}

