% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_storage}
\alias{drill_storage}
\title{Get the list of storage plugin names and configurations}
\usage{
drill_storage(drill_con, plugin = NULL)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{plugin}{the assigned name in the storage plugin definition.}
}
\description{
Get the list of storage plugin names and configurations
}
\examples{
\dontrun{
drill_connection() \%>\% drill_storage()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
