% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_system_reset}
\alias{drill_system_reset}
\title{Changes (optionally, all) system settings back to system defaults}
\usage{
drill_system_reset(drill_con, ..., all = FALSE)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{...}{bare name of system options to reset}

\item{all}{if \code{TRUE}, all parameters are reset (\code{...} is ignored)}
}
\description{
Changes (optionally, all) system settings back to system defaults
}
\examples{
\dontrun{
drill_connection() \%>\% drill_system_reset(all=TRUE)
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

