% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_threads}
\alias{drill_threads}
\title{Get information about threads}
\usage{
drill_threads(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get information about threads
}
\note{
The output of this is in a "viewer" window
}
\examples{
\dontrun{
drill_threads()
}
}

