% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{class}
\name{DrillConnection-class}
\alias{DrillConnection-class}
\title{Class representing a (DBI) database connection which uses JDBC to connect to Drill.}
\description{
Class representing a (DBI) database connection which uses JDBC to connect to Drill.
}

