% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\name{DrillJDBC}
\alias{DrillJDBC}
\title{Drill's JDBC driver main class loader}
\usage{
DrillJDBC()
}
\description{
Drill's JDBC driver main class loader
}
\seealso{
Other Drill JDBC API: \code{\link{dbConnect,DrillJDBCDriver-method}},
  \code{\link{dbDataType,DrillJDBCConnection-method}},
  \code{\link{db_data_type.DrillJDBCConnection}},
  \code{\link{drill_jdbc}}
}
\concept{Drill JDBC API}
