% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{class}
\name{DrillResult-class}
\alias{DrillResult-class}
\title{Representation of a DBI result set returned from a JDBC Drill connection}
\description{
Representation of a DBI result set returned from a JDBC Drill connection
}

