% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\name{format.DrillConnection}
\alias{format.DrillConnection}
\title{A concise character representation (label) for a \code{DrillConnection}}
\usage{
\method{format}{DrillConnection}(x, ...)
}
\arguments{
\item{x}{a \code{DrillConnection}}

\item{...}{ignored}
}
\description{
A concise character representation (label) for a \code{DrillConnection}
}
