% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{print.drill_conn}
\alias{print.drill_conn}
\title{Print function for \code{drill_conn} objects}
\usage{
\method{print}{drill_conn}(x, ...)
}
\arguments{
\item{x}{a \code{drill_conn} object made with \code{\link[=drill_connection]{drill_connection()}}}

\item{...}{unused}
}
\description{
Print function for \code{drill_conn} objects
}
