% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant-package.r
\name{sergeant-exports}
\alias{sergeant-exports}
\alias{tbl}
\title{sergeant exported operators}
\description{
The following functions are imported and then re-exported
from the sergeant package to enable use of the magrittr
pipe operator with no additional library calls
}
