% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill-docker.R
\name{showall_drill}
\alias{showall_drill}
\title{Show all dead and running Drill Docker containers}
\usage{
showall_drill()
}
\description{
This function will show \emph{all} Docker containers that are based on an
image matching a runtime command of "\code{bin/drill-embedded}".
}
\seealso{
Other Drill Docker functions: \code{\link{drill_up}},
  \code{\link{killall_drill}}
}
\concept{Drill Docker functions}
