% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{src_tbls.src_drill}
\alias{src_tbls.src_drill}
\alias{db_desc.src_drill}
\alias{sql_escape_ident.DrillConnection}
\alias{copy_to.src_drill}
\alias{db_explain.DrillConnection}
\alias{db_query_fields.DrillConnection}
\alias{db_data_type.DrillConnection}
\alias{sql_translate_env.DrillConnection}
\title{src tbls}
\usage{
\method{src_tbls}{src_drill}(x, ...)

\method{db_desc}{src_drill}(x)

\method{sql_escape_ident}{DrillConnection}(con, x)

\method{copy_to}{src_drill}(dest, df, name, overwrite, ...)

\method{db_explain}{DrillConnection}(con, sql, ...)

\method{db_query_fields}{DrillConnection}(con, sql, ...)

\method{db_data_type}{DrillConnection}(con, fields, ...)

\method{sql_translate_env}{DrillConnection}(con)
}
\arguments{
\item{x}{x}

\item{...}{ignored}
}
\description{
"SHOW DATABASES"
}
\seealso{
Other Drill REST \code{dplyr} API: 
\code{\link{drill_custom_functions}},
\code{\link{src_drill}()}
}
\concept{Drill REST \code{dplyr} API}
\keyword{internal}
