% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{methods}
\name{dbDataType,DrillConnection-method}
\alias{dbDataType,DrillConnection-method}
\title{Drill dbDataType}
\usage{
\S4method{dbDataType}{DrillConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \code{\linkS4class{DrillDriver}} object}

\item{obj}{Any R object}

\item{...}{Extra optional parameters}
}
\description{
Drill dbDataType
}

