% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbDataType,DrillConnection-method}
\alias{dbDataType,DrillConnection-method}
\title{Drill dbDataType}
\usage{
\S4method{dbDataType}{DrillConnection}(dbObj, obj, ...)
}
\description{
Drill dbDataType
}

