% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{methods}
\name{dbDataType,DrillJDBCConnection-method}
\alias{dbDataType,DrillJDBCConnection-method}
\title{Drill JDBC dbDataType}
\usage{
\S4method{dbDataType}{DrillJDBCConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \code{\linkS4class{DrillJDBCDriver}} object}

\item{obj}{Any R object}

\item{...}{Extra optional parameters}
}
\description{
Drill JDBC dbDataType
}
\seealso{
Other Drill JDBC API: \code{\link{DrillJDBC}},
  \code{\link{dbConnect,DrillJDBCDriver-method}},
  \code{\link{db_data_type.DrillJDBCConnection}},
  \code{\link{drill_jdbc}}
}
\concept{Drill JDBC API}
