% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\name{dbUnloadDriver,DrillDriver-method}
\alias{dbUnloadDriver,DrillDriver-method}
\title{Unload driver}
\usage{
\S4method{dbUnloadDriver}{DrillDriver}(drv, ...)
}
\arguments{
\item{drv}{driver}

\item{...}{Extra optional parameters}
}
\description{
Unload driver
}
\seealso{
Other Drill REST DBI API: 
\code{\link{DrillConnection-class}},
\code{\link{DrillDriver-class}},
\code{\link{DrillResult-class}},
\code{\link{Drill}()},
\code{\link{dbDataType,DrillConnection-method}}
}
\concept{Drill REST DBI API}
