% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{methods}
\name{Drill}
\alias{Drill}
\alias{dbConnect,DrillDriver-method}
\title{Drill}
\usage{
Drill()

\S4method{dbConnect}{DrillDriver}(drv, host = "localhost", port = 8047L,
  ssl = FALSE, username = NULL, password = NULL, ...)
}
\arguments{
\item{drv}{An object created by \code{Drill()}}

\item{host}{host}

\item{port}{port}

\item{ssl}{use ssl?}

\item{username, password}{credentials}

\item{...}{Extra optional parameters}
}
\description{
Drill

Connect to Drill
}
\seealso{
Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}},
  \code{\link{DrillResult-class}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}

Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}},
  \code{\link{DrillResult-class}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}
}
\concept{Drill REST DBI API}
