% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_active}
\alias{drill_active}
\title{Test whether Drill HTTP REST API server is up}
\usage{
drill_active(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
This is a very simple test (performs \code{HEAD /} request on the Drill server/cluster)
}
\examples{
\dontrun{
drill_connection() \%>\% drill_active()
}
}

