% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.r
\name{drill_active}
\alias{drill_active}
\title{Test whether Drill HTTP Drill direct REST API Interface server is up}
\usage{
drill_active(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
This is a very simple test (performs \code{HEAD /} request on the Drill server/cluster)
}
\examples{
try({
drill_connection() \%>\% drill_active()
}, silent=TRUE)
}
\seealso{
Other Drill direct REST API Interface: \code{\link{drill_cancel}},
  \code{\link{drill_connection}},
  \code{\link{drill_functions}},
  \code{\link{drill_metrics}}, \code{\link{drill_options}},
  \code{\link{drill_opts}}, \code{\link{drill_profiles}},
  \code{\link{drill_profile}}, \code{\link{drill_query}},
  \code{\link{drill_settings_reset}},
  \code{\link{drill_set}}, \code{\link{drill_show_files}},
  \code{\link{drill_show_schemas}},
  \code{\link{drill_stats}}, \code{\link{drill_status}},
  \code{\link{drill_storage}},
  \code{\link{drill_system_reset}},
  \code{\link{drill_threads}}, \code{\link{drill_use}},
  \code{\link{drill_version}}
}
\concept{Drill direct REST API Interface}
