% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_cancel}
\alias{drill_cancel}
\title{Cancel the query that has the given queryid.}
\usage{
drill_cancel(query_id, drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{query_id}{the UUID of the query in standard UUID format that Drill assigns to each query.}

\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Cancel the query that has the given queryid.
}

