% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.r
\docType{class}
\name{db_data_type.DrillJDBCConnection}
\alias{db_data_type.DrillJDBCConnection}
\alias{db_data_type.tbl_drill_jdbc}
\alias{DrillJDBCResult-class}
\alias{dbSendQuery,DrillJDBCConnection,character-method}
\alias{sql_escape_ident.DrillJDBCConnection}
\alias{sql_translate_env.DrillJDBCConnection}
\alias{src_tbls.src_dbi}
\title{Drill internals}
\usage{
\method{db_data_type}{DrillJDBCConnection}(con, fields, ...)

\method{db_data_type}{tbl_drill_jdbc}(con, fields, ...)

\S4method{dbSendQuery}{DrillJDBCConnection,character}(conn, statement, ...,
  list = NULL)

\method{sql_escape_ident}{DrillJDBCConnection}(con, x)

\method{sql_translate_env}{DrillJDBCConnection}(con)

\method{src_tbls}{src_dbi}(x)
}
\arguments{
\item{x}{x}
}
\description{
"SHOW DATABASES"
}
\seealso{
Other Drill JDBC API: \code{\link{DrillJDBC}},
  \code{\link{dbConnect,DrillJDBCDriver-method}},
  \code{\link{dbDataType,DrillJDBCConnection-method}},
  \code{\link{drill_jdbc}}
}
\concept{Drill JDBC API}
\keyword{internal}
