% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_metrics}
\alias{drill_metrics}
\title{Get the current memory metrics}
\usage{
drill_metrics(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get the current memory metrics
}
\examples{
\dontrun{
drill_metrics()
}
}

