% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_options}
\alias{drill_options}
\title{List the name, default, and data type of the system and session options}
\usage{
drill_options(drill_con, pattern = NULL)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{pattern}{pattern to filter results by}
}
\description{
List the name, default, and data type of the system and session options
}
\examples{
\dontrun{
drill_connection() \%>\% drill_options()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

