% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.r
\name{drill_options}
\alias{drill_options}
\title{List the name, default, and data type of the system and session options}
\usage{
drill_options(drill_con, pattern = NULL)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{pattern}{pattern to filter results by}
}
\description{
List the name, default, and data type of the system and session options
}
\examples{
\dontrun{
drill_connection() \%>\% drill_options()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
\seealso{
Other Drill direct REST API Interface: 
\code{\link{drill_active}()},
\code{\link{drill_cancel}()},
\code{\link{drill_connection}()},
\code{\link{drill_functions}()},
\code{\link{drill_metrics}()},
\code{\link{drill_opts}()},
\code{\link{drill_profiles}()},
\code{\link{drill_profile}()},
\code{\link{drill_query}()},
\code{\link{drill_settings_reset}()},
\code{\link{drill_set}()},
\code{\link{drill_stats}()},
\code{\link{drill_status}()},
\code{\link{drill_storage}()},
\code{\link{drill_system_reset}()},
\code{\link{drill_threads}()},
\code{\link{drill_version}()}
}
\concept{Drill direct REST API Interface}
