% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_profile}
\alias{drill_profile}
\title{Get the profile of the query that has the given queryid}
\usage{
drill_profile(drill_con, query_id)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{query_id}{UUID of the query in standard UUID format that Drill assigns to each query}
}
\description{
Get the profile of the query that has the given queryid
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

