% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_profile}
\alias{drill_profile}
\title{Get the profile of the query that has the given queryid}
\usage{
drill_profile(drill_con, query_id)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{query_id}{UUID of the query in standard UUID format that Drill assigns to each query}
}
\description{
Get the profile of the query that has the given queryid
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
\seealso{
Other Dill direct REST API Interface: \code{\link{drill_active}},
  \code{\link{drill_cancel}},
  \code{\link{drill_connection}},
  \code{\link{drill_functions}},
  \code{\link{drill_metrics}}, \code{\link{drill_options}},
  \code{\link{drill_opts}}, \code{\link{drill_profiles}},
  \code{\link{drill_query}},
  \code{\link{drill_settings_reset}},
  \code{\link{drill_set}}, \code{\link{drill_show_files}},
  \code{\link{drill_show_schemas}},
  \code{\link{drill_stats}}, \code{\link{drill_status}},
  \code{\link{drill_storage}},
  \code{\link{drill_system_reset}},
  \code{\link{drill_threads}}, \code{\link{drill_use}},
  \code{\link{drill_version}}
}
\concept{Dill direct REST API Interface}
