% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_profiles}
\alias{drill_profiles}
\title{Get the profiles of running and completed queries}
\usage{
drill_profiles(drill_server = Sys.getenv("DRILL_URL"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get the profiles of running and completed queries
}

