% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_settings_reset}
\alias{drill_settings_reset}
\title{Changes (optionally, all) session settings back to system defaults}
\usage{
drill_settings_reset(drill_con, ...)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{...}{bare name of system options to reset}
}
\description{
Changes (optionally, all) session settings back to system defaults
}
\examples{
\dontrun{
drill_connection() \%>\% drill_settings_reset(exec.errors.verbose)
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
\seealso{
Other Drill direct REST API Interface: \code{\link{drill_active}},
  \code{\link{drill_cancel}},
  \code{\link{drill_connection}},
  \code{\link{drill_functions}},
  \code{\link{drill_metrics}}, \code{\link{drill_options}},
  \code{\link{drill_opts}}, \code{\link{drill_profiles}},
  \code{\link{drill_profile}}, \code{\link{drill_query}},
  \code{\link{drill_set}}, \code{\link{drill_stats}},
  \code{\link{drill_status}}, \code{\link{drill_storage}},
  \code{\link{drill_system_reset}},
  \code{\link{drill_threads}}, \code{\link{drill_version}}
}
\concept{Drill direct REST API Interface}
