% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_settings_reset}
\alias{drill_settings_reset}
\title{Changes (optionally, all) session settings back to system defaults}
\usage{
drill_settings_reset(drill_con, ...)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{...}{bare name of system options to reset}
}
\description{
Changes (optionally, all) session settings back to system defaults
}
\examples{
\dontrun{
drill_connection() \%>\% drill_settings_reset(exec.errors.verbose)
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

