% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{drill_show_files}
\alias{drill_show_files}
\title{Show files in a file system schema.}
\usage{
drill_show_files(drill_con, schema_spec)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{schema_spec}{properly quoted "filesystem.directory_name" reference path}
}
\description{
Show files in a file system schema.
}
\examples{
\dontrun{
drill_connection() \%>\% drill_show_files("dfs.tmp")
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
