% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{drill_show_schemas}
\alias{drill_show_schemas}
\title{Returns a list of available schemas.}
\usage{
drill_show_schemas(drill_con, .progress = interactive())
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{.progress}{if \code{TRUE} (default if in an interactive session) then ask
\code{httr::POST} to display a progress bar}
}
\description{
Returns a list of available schemas.
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
\seealso{
Other Dill direct REST API Interface: 
\code{\link{drill_show_files}()},
\code{\link{drill_use}()}
}
\concept{Dill direct REST API Interface}
