% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{drill_show_schemas}
\alias{drill_show_schemas}
\title{Returns a list of available schemas.}
\usage{
drill_show_schemas(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Returns a list of available schemas.
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

