% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_stats}
\alias{drill_stats}
\title{Get Drillbit information, such as ports numbers}
\usage{
drill_stats(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Get Drillbit information, such as ports numbers
}
\examples{
\dontrun{
drill_con() \%>\% drill_stats()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

