% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_status}
\alias{drill_status}
\title{Get the status of Drill}
\usage{
drill_status(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Get the status of Drill
}
\note{
The output of this is in a "viewer" window
}
\examples{
\dontrun{
drill_connection() \%>\% drill_status()
}
}
