% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_storage}
\alias{drill_storage}
\title{Get the list of storage plugin names and configurations}
\usage{
drill_storage(drill_server = Sys.getenv("DRILL_URL"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Get the list of storage plugin names and configurations
}

