% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{drill_system_reset}
\alias{drill_system_reset}
\title{Changes (optionally, all) system settings back to system defaults}
\usage{
drill_system_reset(drill_con, ..., all = FALSE)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{...}{bare name of system options to reset}

\item{all}{if \code{TRUE}, all parameters are reset (\code{...} is ignored)}
}
\description{
Changes (optionally, all) system settings back to system defaults
}
\examples{
\dontrun{
drill_connection() \%>\% drill_system_reset(all=TRUE)
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
\seealso{
Other Dill direct REST API Interface: \code{\link{drill_active}},
  \code{\link{drill_cancel}},
  \code{\link{drill_connection}},
  \code{\link{drill_metrics}}, \code{\link{drill_options}},
  \code{\link{drill_profiles}},
  \code{\link{drill_profile}}, \code{\link{drill_query}},
  \code{\link{drill_settings_reset}},
  \code{\link{drill_set}}, \code{\link{drill_show_files}},
  \code{\link{drill_show_schemas}},
  \code{\link{drill_stats}}, \code{\link{drill_status}},
  \code{\link{drill_storage}}, \code{\link{drill_threads}},
  \code{\link{drill_use}}, \code{\link{drill_version}}
}
\concept{Dill direct REST API Interface}
