% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_threads}
\alias{drill_threads}
\title{Get information about threads}
\usage{
drill_threads(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Get information about threads
}
\note{
The output of this is in a "viewer" window
}
\examples{
\dontrun{
drill_connection() \%>\% drill_threads()
}
}
