% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{drill_uplift}
\alias{drill_uplift}
\title{Turn a columnar query results into a type-converted tbl}
\usage{
drill_uplift(query_result)
}
\arguments{
\item{query_result}{the result of a call to `drill_query()`}
}
\description{
If you know the result of `drill_query()` will be a data frame, then
you can pipe it to this function to pull out `rows` and automatically
type-convert it.
}
\details{
Not really intended to be called directly, but useful if you ran \code{drill_query()}
without `uplift=TRUE` but want to then convert the structure.
}

