% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{drill_use}
\alias{drill_use}
\title{Change to a particular schema.}
\usage{
drill_use(schema_name, drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{schema_name}{A unique name for a Drill schema. A schema in Drill is a configured
storage plugin, such as hive, or a storage plugin and workspace.}

\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Change to a particular schema.
}

