% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{drill_use}
\alias{drill_use}
\title{Change to a particular schema.}
\usage{
drill_use(drill_con, schema_name)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}

\item{schema_name}{A unique name for a Drill schema. A schema in Drill is a configured
storage plugin, such as hive, or a storage plugin and workspace.}
}
\description{
Change to a particular schema.
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}
