% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_version}
\alias{drill_version}
\title{Identify the version of Drill running}
\usage{
drill_version(drill_con)
}
\arguments{
\item{drill_con}{drill server connection object setup by \code{drill_connection()}}
}
\description{
Identify the version of Drill running
}
\examples{
\dontrun{
drill_connection() \%>\% drill_version()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

