% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant.r
\name{drill_version}
\alias{drill_version}
\title{Identify the version of Drill running}
\usage{
drill_version(drill_server = Sys.getenv("DRILL_URL", unset =
  "http://localhost:8047"))
}
\arguments{
\item{drill_server}{base URL of the \code{drill} server}
}
\description{
Identify the version of Drill running
}
\examples{
\dontrun{
drill_version()
}
}
\references{
\href{https://drill.apache.org/docs/}{Drill documentation}
}

