% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{class}
\name{DrillConnection-class}
\alias{DrillConnection-class}
\alias{dbSendQuery,DrillConnection,ANY-method}
\alias{dbSendQuery,DrillConnection,character-method}
\alias{dbIsValid,DrillConnection-method}
\alias{dbListFields,DrillConnection,character-method}
\title{Drill connection class.}
\usage{
\S4method{dbSendQuery}{DrillConnection,ANY}(conn, statement, ...)

\S4method{dbIsValid}{DrillConnection}(dbObj, ...)

\S4method{dbListFields}{DrillConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{connection}

\item{statement}{SQL statement}

\item{...}{passed on to methods}
}
\description{
Drill connection class.

Send a query to Drill
}
\keyword{internal}
