% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{class}
\name{DrillDriver-class}
\alias{DrillDriver-class}
\title{Driver for Drill database.}
\description{
Driver for Drill database.
}
\seealso{
Other Drill REST DBI API: 
\code{\link{DrillConnection-class}},
\code{\link{DrillResult-class}},
\code{\link{Drill}()},
\code{\link{dbDataType,DrillConnection-method}},
\code{\link{dbUnloadDriver,DrillDriver-method}}
}
\concept{Drill REST DBI API}
\keyword{internal}
