% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\docType{class}
\name{DrillResult-class}
\alias{DrillResult-class}
\alias{dbClearResult,DrillResult-method}
\alias{dbFetch,DrillResult-method}
\alias{dbHasCompleted,DrillResult-method}
\alias{dbListFields,DrillResult,missing-method}
\alias{dbGetStatement,DrillResult-method}
\title{Drill results class.}
\usage{
\S4method{dbClearResult}{DrillResult}(res, ...)

\S4method{dbFetch}{DrillResult}(res, .progress = FALSE, ...)

\S4method{dbHasCompleted}{DrillResult}(res, ...)

\S4method{dbListFields}{DrillResult,missing}(conn, name)

\S4method{dbGetStatement}{DrillResult}(res, ...)
}
\arguments{
\item{.progress}{show data transfer progress?}
}
\description{
Drill results class.

Clear

Retrieve records from Drill query

Completed

Statement
}
\seealso{
Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}}, \code{\link{Drill}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}

Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}}, \code{\link{Drill}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}

Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}}, \code{\link{Drill}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}

Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}}, \code{\link{Drill}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}

Other Drill REST DBI API: \code{\link{DrillConnection-class}},
  \code{\link{DrillDriver-class}}, \code{\link{Drill}},
  \code{\link{dbDataType,DrillConnection-method}},
  \code{\link{dbUnloadDriver,DrillDriver-method}}
}
\concept{Drill REST DBI API}
\keyword{internal}
