% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergeant-package.r
\docType{package}
\name{sergeant}
\alias{sergeant}
\alias{sergeant-package}
\title{Tools to Transform and Query Data with the 'Apache' 'Drill' 'API'}
\description{
Tools to Transform and Query Data with the 'Apache' 'Drill' 'API'
}
\author{
Bob Rudis (bob@rud.is)
}

